design-projects/micro-services-with-open-telemetry/gateway/requirements.txt
fastapi
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
requests
psycopg2-binary
opentelemetry-instrumentation-requests
